+++ CHANGELOG.md
@@ -45,6 +45,7 @@ Types of changes:
 - GitHub action to add a changelog reminder
 - View document on mobile before sign
 - Markdown formatting for description
+- Legal information on Validation screen, configure in Admin settings
 - Button to redirect to files to view the document
 
 # Changed